Product Controller

+++ routes/web.php
@@ -8,9 +8,9 @@ Route::group([
     'prefix' => '/product'
 ] , function(){
 
-   Route::get('/' , [ProductController::class , 'all']);
+   Route::get('/' , [ProductController::class , 'all'])->name('admin.product.all');
    Route::get('/add' , [ProductController::class , 'add']);
-   Route::post('/add' , [ProductController::class , 'store']);
-   Route::delete('/delete/{id}' , [ProductController::class , 'delete']);
+   Route::post('/add' , [ProductController::class , 'store'])->name('admin.product.store');
+   Route::delete('/delete/{id}' , [ProductController::class , 'delete'])->name('admin.product.delete');
 
 });
